Update default view, still need better iconography, maybe more tweks to colors, and on selection of a specific line, visually highlight the corresponding metric in the detail view

+++ WeatherThereIsNewsOrNot/Weather/Weather/Model/WeatherSnapshot.swift
@@ -19,7 +19,7 @@ struct WeatherSnapshot: Unmarshaling {
     
     var time: Int
     var summary: String
-    var precipProbability: Int
+    var precipProbability: Float
     var temperature: Float
     var humidity: Float
     var windSpeed: Float
